src/requirements.txt
alembic==1.9.1
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.0
Faker==15.3.4
fastapi==0.89.1
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.1
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
pydantic==1.10.4
pytest==7.2.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tenacity==8.1.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
